feat(desktop): W7.3 extract DesktopResponder (impl AgentResponder) from ChatDelegate::call_llm
ADR-160 §3 L2 W7.3 — extract LLM call pipeline (cost guard, per-user
selected_model override, streaming/non-streaming dispatch, cost record,
cache monitor, DB persistence) out of ChatDelegate::call_llm into a new
DesktopResponder type that implements dasclaw_runtime::AgentResponder.

ChatDelegate::call_llm becomes a one-line delegation. The iteration==0
gate for selected_model override is replaced by an AtomicBool sentinel
because AgentResponder::respond doesn't receive an iteration param.

Refs #967
