+++ desktop-client/ironclaw/src/agent/mod.rs
@@ -16,6 +16,7 @@ mod attachments;
 mod commands;
 pub mod compaction;
 mod desktop_dispatcher;
+mod desktop_responder;
 mod dispatcher;
 mod router;
 pub mod session;
